Perform some refactorings
The way worker nodes join to the cluster (for testing purposes) does
not resemble a production environment, because they cannot go through
the "external interface" of the public hypervisor, and thus, they are
connecting to the API server through the "internal interface" of the
public hypervisor.

This might cause false positives and wrong results when we do more
complex networking.