Initial implementation of worker node join
This allows `oi node join` to create a NodeJoinRequest resource, that
contains a public key. This NodeJoinRequest resource will be
reconciled by the main oneinfra controller/binary, allocating a new
VPN entry for this new node, and updating the NodeJoinRequest status
subresource.

This is just the initial implementation, in a working status. However,
the whole strategy for joining the node itself is not yet done, it is
required to:

- Create the kubeconfig file for the new node

- Create the kubelet config file for the new node

- Make the `oi node join` wait until everything has been set up by the
  main oneinfra controllers, or timeout. Or add a new command to
  explicitly wait for the `issued` condition of the node join request
  resource.

- Allow to create a cloud-init file that will contain all this logic,
  that you can just leverage to your cloud provider

- Create a worker image that will contain everything needed as
  dependencies to perform all these tasks

Fixes https://github.com/oneinfra/oneinfra/issues/3