CLI: add ability to retry reconciliation
Since the reconciliation process is meant to be executed in an endless
loop it can go through some error transient states, that will be
amended on the following reconcile cycles. However, when the CLI is
used, a different strategy needs to be adopted, since it's not an
endless loop.

Thus, include the `max-retries` and `retry-wait-time` arguments to the
`oi reconcile` CLI. If you are using the CLI interface, this will
allow you to set how many `max-retries` you want to run if the
reconciliation process failed, and the `retry-wait-time` between
global reconciliation loop retries.