`oi-local-cluster`: pub and priv hypervisors have the same semantics
It makes no sense to have different semantics locally for public and
private hypervisors. It will only make it harder to for example, join
new test fake workers, because the internal IP address of public
hypervisors was set to 127.0.0.1 (with port exposed on the host), so
when it tried to use the Docker IP, the certificate for the apiserver
was invalid.

Since `oi-local-cluster` is only for testing locally (and not trying
to resemble a production environment that close), it's fine if both
public and private hypervisors have the same networking semantics. For
the purpose of trying it out, scheduling will respect public and
private hypervisors though, as expected.