Publish join information in a ConfigMap
Join requests now will use the bootstrap token to also read the public
key of the cluster to be joined, so they can cipher the generated
symmetric key with it. There is no need to pass the cluster join
public key to the new joiner, since it will validate the CA of the
apiserver and then download the public key from it.

Fixes: https://github.com/oneinfra/oneinfra/issues/33