cluster mutating webhook: simplify and optimize
Cluster mutating webhooks should not generate certificates, since it
will block API calls on the API server.

Instead, only default trivial values, and let the reconcile loop take
care of generating the certificates if required.

This should be improved yet: the mutating webhook should annotate the
cluster with a label if some certificate or key is missing, and a
separate controller should be watching clusters with these labels.

Fixes: https://github.com/oneinfra/oneinfra/issues/36