Add `LastSetTime` to conditions
We had `LastTransitionTime` for conditions, that record when was the
last time a condition did transition. It is good to keep this record
of when a condition did change the last time.

At the same time, it would be interesting to know when was the last
time a condition was set, even if it didn't transition at all. This
will be the `LastSetTime` duty.

Whenever a condition is set, `LastSetTime` will contain the timestamp
of when that happened. `LastTransitionTime` will still only report the
latest transition time.