Remove etcd members with some level of retries
Errors should be differentiated, and have different strategies for
example for a returned error of "unhealthy server" by the etcd
cluster, or if a context deadline happened.

For now, blindly retrying for a period of time will allow us to make
sure the etcd member is removed in a reasonable amount of time, or
bail out.