Add the notion of input and output endpoints
This allows for every component to store in its status its intent:
whether it takes endpoints as input (and which), and whether it
outputs any number of endpoints (and which).

This is specially useful when reconciling some components because of a
side effect. For example, when a Cluster is scaled down, and
components removed, the ingress should be reconciled in order to point
to the remaining components; similar when scaling up.

Fixes: https://github.com/oneinfra/oneinfra/issues/50