apiserver: register correct kubernetes apiserver endpoint
When the apiservers register the kubernetes service endpoints, they
will use their advertise address and the bind port.

This sets the proper advertise address to the ingress hypervisor IP
address, and tricks the apiserver in listening in the port that will
be assigned to the ingress hypervisor port, forwarded in private
hypervisors to the assigned port.