Fix fake hypervisor creation: use kind network if exists
If no fake hypervisor network is provided (new argument), check if a
`kind` network exists and use that.

Since kind started to create its own bridge network (by default
`kind`), starting with kind 0.8.0 it was impossible for the controller
manager to contact the fake hypervisors on the remote CRI endpoints.

Now, it's possible to configure the docker network of the fake
hypervisor set, or if omitted, `oi-local-hypervisor-set` will check
whether a `kind` docker network exists. If no network is provided, and
this docker network does not exist, the old behavior will be
respected (the containers are not attached to any specific network).

Fixes: https://github.com/oneinfra/oneinfra/issues/57